+++ crates/rust-client/src/sync/mod.rs
@@ -54,17 +54,16 @@
 //! `committed_note_updates` and `consumed_note_updates`) to understand how the sync data is
 //! processed and applied to the local store.
 
-use alloc::{boxed::Box, vec::Vec};
+use alloc::vec::Vec;
 use core::cmp::max;
 
 use miden_objects::{
     account::AccountId,
     block::BlockNumber,
     crypto::rand::FeltRng,
-    note::{NoteId, NoteTag, Nullifier},
+    note::{NoteId, NoteTag},
     transaction::TransactionId,
 };
-use state_sync::on_note_received;
 
 use crate::{store::NoteFilter, Client, ClientError};
 
@@ -74,7 +73,7 @@ mod tag;
 pub use tag::{NoteTagRecord, NoteTagSource};
 
 mod state_sync;
-pub use state_sync::{OnNoteReceived, OnNullifierReceived, StateSync};
+pub use state_sync::{on_note_received, OnNoteReceived, StateSync};
 
 mod state_sync_update;
 pub use state_sync_update::StateSyncUpdate;
@@ -82,10 +81,7 @@ pub use state_sync_update::StateSyncUpdate;
 // CONSTANTS
 // ================================================================================================
 
-/// The number of bits to shift identifiers for in use of filters.
-pub(crate) const FILTER_ID_SHIFT: u8 = 48;
-
-/// Client syncronization methods.
+// Client syncronization methods.
 impl<R: FeltRng> Client<R> {
     // SYNC STATE
     // --------------------------------------------------------------------------------------------
@@ -117,17 +113,6 @@ impl<R: FeltRng> Client<R> {
     pub async fn sync_state(&mut self) -> Result<SyncSummary, ClientError> {
         _ = self.ensure_genesis_in_place().await?;
 
-        let state_sync = StateSync::new(
-            self.rpc_api.clone(),
-            Box::new({
-                let store_clone = self.store.clone();
-                move |committed_note, public_note| {
-                    Box::pin(on_note_received(store_clone.clone(), committed_note, public_note))
-                }
-            }),
-            Box::new(move |_committed_note| Box::pin(async { Ok(true) })),
-        );
-
         // Get current state of the client
         let accounts = self
             .store
@@ -144,7 +129,8 @@ impl<R: FeltRng> Client<R> {
         let unspent_output_notes = self.store.get_output_notes(NoteFilter::Unspent).await?;
 
         // Get the sync update from the network
-        let state_sync_update = state_sync
+        let state_sync_update = self
+            .state_sync_component
             .sync_state(
                 self.build_current_partial_mmr().await?,
                 accounts,
@@ -166,10 +152,6 @@ impl<R: FeltRng> Client<R> {
 
         Ok(sync_summary)
     }
-}
-
-pub(crate) fn get_nullifier_prefix(nullifier: &Nullifier) -> u16 {
-    (nullifier.inner()[3].as_int() >> FILTER_ID_SHIFT) as u16
 }
 
 // SYNC SUMMARY