+++ tests/integration/common.rs
@@ -14,7 +14,7 @@ use miden_client::{
     note::create_p2id_note,
     rpc::{Endpoint, RpcError, TonicRpcClient},
     store::{sqlite_store::SqliteStore, NoteFilter, TransactionFilter},
-    sync::SyncSummary,
+    sync::{on_note_received, StateSync, SyncSummary},
     testing::account_id::ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_OFF_CHAIN,
     transaction::{
         DataStoreError, TransactionExecutorError, TransactionRequest, TransactionRequestBuilder,
@@ -61,16 +61,22 @@ pub async fn create_test_client() -> (TestClient, FilesystemKeyStore) {
     let rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
 
     let keystore = FilesystemKeyStore::new(auth_path).unwrap();
-
     let authenticator = ClientAuthenticator::new(rng, keystore.clone());
+
+    let rpc_api = Arc::new(TonicRpcClient::new(&rpc_endpoint, rpc_timeout));
+
+    let state_sync_component = StateSync::new(
+        rpc_api.clone(),
+        Box::new({
+            let store_clone = store.clone();
+            move |committed_note, public_note| {
+                Box::pin(on_note_received(store_clone.clone(), committed_note, public_note))
+            }
+        }),
+    );
+
     (
-        TestClient::new(
-            Arc::new(TonicRpcClient::new(&rpc_endpoint, rpc_timeout)),
-            rng,
-            store,
-            Arc::new(authenticator),
-            true,
-        ),
+        TestClient::new(rpc_api, rng, store, Arc::new(authenticator), state_sync_component, true),
         keystore,
     )
 }
@@ -216,22 +222,15 @@ pub async fn execute_tx_and_sync(
 pub async fn wait_for_tx(client: &mut TestClient, transaction_id: TransactionId) {
     // wait until tx is committed
     loop {
-        println!("Syncing State...");
-        client.sync_state().await.unwrap();
+        let summary = client.sync_state().await.unwrap();
+        println!("Syncing State {}...", summary.block_num.as_u32());
 
-        // Check if executed transaction got committed by the node
-        let uncommited_transactions =
-            client.get_transactions(TransactionFilter::Uncomitted).await.unwrap();
-        let is_tx_committed = uncommited_transactions
-            .iter()
-            .all(|uncommited_tx| uncommited_tx.id != transaction_id);
-
-        if is_tx_committed {
+        if summary.committed_transactions.contains(&transaction_id) {
             break;
         }
 
         // 500_000_000 ns = 0.5s
-        std::thread::sleep(std::time::Duration::new(0, 500_000_000));
+        tokio::time::sleep(std::time::Duration::new(0, 500_000_000)).await;
     }
 }